backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "opencv-python-headless>=4.13.0.90",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "ultralytics-opencv-headless>=8.4.6",
    "uvicorn>=0.40.0",
]
